Привел к однотипности контроллеры, сделал все численные данные типа Long

+++ employees-service/src/main/java/com/example/employees_service/dto/EmployeeCreateUpdateDto.java
@@ -20,7 +20,7 @@ public class EmployeeCreateUpdateDto {
 
     private String patronymic;
 
-    private int stuffId;
+    private Long stuffId;
 
     private String employeePost;
